feqw

+++ Kernel/keyboard.c
@@ -161,5 +161,5 @@ uint8_t getLastInput()
 	}
 	char aux;
 	fd_read(KEYBOARD, &aux, 1);
-	return aux;
+	return (uint8_t)aux;
 }
